+++ src/com/tagbio/umap/metric/EuclideanMetric.java
@@ -17,13 +17,13 @@ public class EuclideanMetric extends Metric {
   }
 
   @Override
-  public double distance(final float[] x, final float[] y) {
+  public float distance(final float[] x, final float[] y) {
     //  D(x, y) = \sqrt{\sum_i (x_i - y_i)^2}
     float result = 0;
     for (int i = 0; i < x.length; ++i) {
       final float d = x[i] - y[i];
       result += d * d;
     }
-    return Math.sqrt(result);
+    return (float) Math.sqrt(result);
   }
 }
