+++ src/com/tagbio/umap/metric/MinkowskiMetric.java
@@ -18,12 +18,12 @@ public class MinkowskiMetric extends Metric {
   }
 
   @Override
-  public double distance(final float[] x, final float[] y) {
+  public float distance(final float[] x, final float[] y) {
     // D(x, y) = \left(\sum_i |x_i - y_i|^p\right)^{\frac{1}{p}}
     double result = 0.0;
     for (int i = 0; i < x.length; ++i) {
       result += Math.pow(Math.abs(x[i] - y[i]), mPower);
     }
-    return Math.pow(result, 1 / mPower);
+    return (float) Math.pow(result, 1 / mPower);
   }
 }
