+++ src/com/tagbio/umap/metric/PrecomputedMetric.java
@@ -18,7 +18,7 @@ public class PrecomputedMetric extends Metric {
   }
 
   @Override
-  public double distance(final float[] x, final float[] y) {
+  public float distance(final float[] x, final float[] y) {
     throw new IllegalStateException("Attempt to computed distance when distances precomputed");
   }
 }
